♻️ refactor: doman -> domain
