+++ src/main/java/naeilmolae/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package naeilmolae.domain.member.repository;
 
-import naeilmolae.domain.member.doman.Member;
+import naeilmolae.domain.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
